Updates to ludown to match other tool's cmd structure

+++ Ludown/lib/parser.js
@@ -14,14 +14,43 @@ module.exports = {
     /**
      * handle parsing the root file that was passed in command line args
      *
-     * @param {string} rootFile Path to the root file passed in command line
      * @param {object} program content flushed out by commander
      */
-    handleFile(rootFile, program) {
+    handleFile(program, cmd) {
         try
         {
+            var filesToParse = [];
+            var rootFile = '';
+
+            if(program.in) {
+                rootFile = program.in;
+                filesToParse.push(program.in);
+
+            }
+            
+            if(program.lu_folder) {
+                // is this a folder? 
+                if(!fs.statSync(program.lu_folder).isDirectory()) {
+                    console.error(chalkdefault.redBright('Sorry, ' + program.lu_folder + ' is not a folder or does not exist'));
+                    process.exit(1);
+                }
+                if(program.subfolder) {
+                    filesToParse = findLUFiles(program.lu_folder, true); 
+                } else {
+                    filesToParse = findLUFiles(program.lu_folder, false); 
+                }
+
+                if(filesToParse.length === 0) {
+                    console.error(chalkdefault.redBright('Sorry, no .lu files found in the specified folder.'));
+                    process.exit(1);
+                }
+                if(!rootFile) rootFile = filesToParse[0]
+            }
+
+            
+
+
             // handle root file and subseqntly own calling parse on other files found in rootFile
-            var filesToParse = [rootFile];
             var allParsedLUISContent = new Array();
             var allParsedQnAContent = new Array();
             
@@ -34,25 +63,25 @@ module.exports = {
                     outFolder = path.resolve('', program.out_folder);
                 }
                 if(!fs.existsSync(outFolder)) {
-                    process.stdout.write(chalk.red('\nOutput folder ' + outFolder + ' does not exist\n'));
+                    process.stdout.write(chalk.default.redBright('\nOutput folder ' + outFolder + ' does not exist\n'));
                     process.exit(1);
                 }
             }
             while(filesToParse.length > 0) {
                 var file = filesToParse[0];
                 if(!fs.existsSync(path.resolve(file))) {
-                    process.stdout.write(chalk.red('Sorry unable to open [' + file + ']\n'));        
+                    process.stdout.write(chalk.default.redBright('Sorry unable to open [' + file + ']\n'));        
                     process.exit(1);
                 }
                 var fileContent = fs.readFileSync(file,'utf8');
                 if (!fileContent) {
-                    process.stdout.write(chalk.red('Sorry, error reading file:' + file + '\n'));    
+                    process.stdout.write(chalk.default.redBright('Sorry, error reading file:' + file + '\n'));    
                     process.exit(1);
                 }
-                if(!program.quiet) process.stdout.write(chalk.cyan('Parsing file: ' + file + '\n'));
-                var parsedContent = parseFileContents.parseFile(fileContent, program.quiet);
+                console.log(chalk.cyan('Parsing file: ' + file + '\n'));
+                var parsedContent = parseFileContents.parseFile(fileContent, program.verbose);
                 if (!parsedContent) {
-                    process.stdout.write(chalk.red('Sorry, file ' + file + 'had invalid content\n'));
+                    process.stdout.write(chalk.default.redBright('Sorry, file ' + file + 'had invalid content\n'));
                     process.exit(1);
                 } else {
                     allParsedLUISContent.push(parsedContent.LUISBlob);
@@ -100,18 +129,12 @@ module.exports = {
                                 (finalLUISJSON.prebuiltEntities.length > 0) ||
                                 (finalLUISJSON.model_features.length > 0);
 
-            if(!program.quiet) {
-                if(writeLUISFile) {
-                    process.stdout.write(chalk.gray('-----------------------------------\n'));
-                    process.stdout.write(chalk.gray('|         FINAL LUIS JSON         |\n'));
-                    process.stdout.write(chalk.gray('-----------------------------------\n'));
-                    process.stdout.write(chalk.gray(JSON.stringify(finalLUISJSON, null, 2) + '\n'));
+            if(program.verbose) {
+                if((cmd == 'luis') && writeLUISFile) {
+                    process.stdout.write(JSON.stringify(finalLUISJSON, null, 2) + '\n');
                 }
-                if(writeQnAFile) {
-                    process.stdout.write(chalk.gray('-----------------------------------\n'));
-                    process.stdout.write(chalk.gray('|         FINAL QnA JSON          |\n'));
-                    process.stdout.write(chalk.gray('-----------------------------------\n'));
-                    process.stdout.write(chalk.gray(JSON.stringify(finalQnAJSON, null, 2) + '\n'));
+                if((cmd == 'qna') && writeQnAFile) {
+                    process.stdout.write(JSON.stringify(finalQnAJSON, null, 2) + '\n');
                 }
             }
             
@@ -137,26 +160,26 @@ module.exports = {
                     program.qTSVFile = program.qna_name + "_qnaTSV.tsv";
                 }
             }
-            if(writeLUISFile) {
+            if((cmd == 'luis') && writeLUISFile) {
                 // write out the final LUIS Json
                 fs.writeFileSync(outFolder + '\\' + program.lOutFile, JSON.stringify(finalLUISJSON, null, 2), function(error) {
                     if(error) {
-                        process.stdout.write(chalk.red('Unable to write LUIS JSON file - ' + outFolder + '\\' + program.lOutFile + '\n'));
+                        process.stdout.write(chalk.default.redBright('Unable to write LUIS JSON file - ' + outFolder + '\\' + program.lOutFile + '\n'));
                         process.exit(1);
                     } 
                 });
-                if(!program.quiet) process.stdout.write(chalk.green('Successfully wrote LUIS model to ' + outFolder + '\\' + program.lOutFile + '\n'));
+                console.log(chalk.green('Successfully wrote LUIS model to ' + outFolder + '\\' + program.lOutFile + '\n'));
             }
 
-            if(writeQnAFile) {
+            if((cmd == 'qna') && writeQnAFile) {
                 // write out the final LUIS Json
                 fs.writeFileSync(outFolder + '\\' + program.qOutFile, JSON.stringify(finalQnAJSON, null, 2), function(error) {
                     if(error) {
-                        process.stdout.write(chalk.red('Unable to write QnA JSON file - ' + outFolder + '\\' + program.qOutFile + '\n'));
+                        process.stdout.write(chalk.default.redBright('Unable to write QnA JSON file - ' + outFolder + '\\' + program.qOutFile + '\n'));
                         process.exit(1);
                     } 
                 });
-                if(!program.quiet) process.stdout.write(chalk.green('Successfully wrote QnA KB to ' + outFolder + '\\' + program.qOutFile + '\n'));
+                console.log(chalk.green('Successfully wrote QnA KB to ' + outFolder + '\\' + program.qOutFile + '\n'));
 
                 // write tsv file for QnA maker
                 var QnAFileContent = "";
@@ -165,24 +188,24 @@ module.exports = {
                 });
                 fs.writeFileSync(outFolder + '\\' + program.qTSVFile, QnAFileContent, function(error) {
                     if(error) {
-                        process.stdout.write(chalk.red('Unable to write QnA TSV file - ' + outFolder + '\\' + program.qTSVFile + '\n'));
+                        process.stdout.write(chalk.default.redBright('Unable to write QnA TSV file - ' + outFolder + '\\' + program.qTSVFile + '\n'));
                         process.exit(1);
                     } 
                 });
-                if(!program.quiet) process.stdout.write(chalk.green('Successfully wrote QnA TSV to ' + outFolder + '\\' + program.qTSVFile + '\n'));
+                console.log(chalk.green('Successfully wrote QnA TSV to ' + outFolder + '\\' + program.qTSVFile + '\n'));
             }
 
             // write luis batch test file if requested
-            if(program.write_luis_batch_tests) {
+            if((cmd == 'luis') && program.write_luis_batch_tests) {
                 var LUISBatchFileName = program.lOutFile.replace("_LUISApp.json","_LUISBatchTest.json");
                 // write out the final LUIS Json
                 fs.writeFileSync(outFolder + '\\' + LUISBatchFileName, JSON.stringify(finalLUISJSON.utterances, null, 2), function(error) {
                     if(error) {
-                        process.stdout.write(chalk.red('Unable to write LUIS batch test JSON file - ' + outFolder + '\\' + LUISBatchFileName + '\n'));
+                        process.stdout.write(chalk.default.redBright('Unable to write LUIS batch test JSON file - ' + outFolder + '\\' + LUISBatchFileName + '\n'));
                         process.exit(1);
                     } 
                 });
-                if(!program.quiet) process.stdout.write(chalk.green('Successfully wrote LUIS batch test JSON file to ' + outFolder + '\\' +  LUISBatchFileName + '\n'));
+                console.log(chalk.green('Successfully wrote LUIS batch test JSON file to ' + outFolder + '\\' +  LUISBatchFileName + '\n'));
             }
             // write luis only to stdout
             if(program.gen_luis_only) {
@@ -196,13 +219,38 @@ module.exports = {
             }
             process.exit(0);
         } catch (err) {
-            process.stdout.write(chalk.red('Oops! Something went wrong.\n'));
+            process.stdout.write(chalk.default.redBright('Oops! Something went wrong.\n'));
             process.stdout.write(chalk.yellow(err));
             process.exit(1);
         }
     }
 };
+/**
+ * Helper function to recursively get all .lu files
+ * @param {string} inputfolder input folder name
+ * @param {boolean} getSubFolder indicates if we should recursively look in sub-folders as well
+ * @returns {Array} Array of .lu files found
+ */
+var findLUFiles = function(inputFolder, getSubFolders) {
+    var results = [];
+    const luExt = '.lu';
 
+    fs.readdirSync(inputFolder).forEach(function(dirContent) {
+        dirContent = path.resolve(inputFolder,dirContent);
+
+        if(getSubFolders && fs.statSync(dirContent).isDirectory()) {
+            results = results.concat(findLUFiles(dirContent, getSubFolders));
+        }
+
+        if(fs.statSync(dirContent).isFile()) {
+            if(dirContent.endsWith(luExt)) {
+                results.push(dirContent);
+            }
+        }
+    });
+    
+    return results;
+}
 /**
  * Handle collating all QnA sections across all parsed files into one QnA collection
  *
